+++ src/test/java/com/google/firebase/cloud/FirestoreClientTest.java
@@ -7,6 +7,7 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import com.google.auth.oauth2.GoogleCredentials;
+import com.google.cloud.firestore.DocumentReference;
 import com.google.cloud.firestore.Firestore;
 import com.google.cloud.firestore.FirestoreOptions;
 import com.google.firebase.FirebaseApp;
@@ -14,12 +15,20 @@ import com.google.firebase.FirebaseOptions;
 import com.google.firebase.FirebaseOptions.Builder;
 import com.google.firebase.ImplFirebaseTrampolines;
 import com.google.firebase.TestOnlyImplFirebaseTrampolines;
+import com.google.firebase.auth.MockGoogleCredentials;
 import com.google.firebase.testing.ServiceAccount;
 import java.io.IOException;
 import org.junit.After;
 import org.junit.Test;
 
 public class FirestoreClientTest {
+
+  public static final FirestoreOptions FIRESTORE_OPTIONS = FirestoreOptions.newBuilder()
+      // Setting credentials is not required (they get overridden by Admin SDK), but without
+      // this Firestore logs an ugly warning during tests.
+      .setCredentials(new MockGoogleCredentials("test-token"))
+      .setTimestampsInSnapshotsEnabled(true)
+      .build();
 
   @After
   public void tearDown() {
@@ -31,9 +40,7 @@ public class FirestoreClientTest {
     FirebaseApp app = FirebaseApp.initializeApp(new FirebaseOptions.Builder()
         .setCredentials(GoogleCredentials.fromStream(ServiceAccount.EDITOR.asStream()))
         .setProjectId("explicit-project-id")
-        .setFirestoreOptions(FirestoreOptions.newBuilder()
-            .setTimestampsInSnapshotsEnabled(true)
-            .build())
+        .setFirestoreOptions(FIRESTORE_OPTIONS)
         .build());
     Firestore firestore = FirestoreClient.getFirestore(app);
     assertEquals("explicit-project-id", firestore.getOptions().getProjectId());
@@ -46,9 +53,7 @@ public class FirestoreClientTest {
   public void testServiceAccountProjectId() throws IOException {
     FirebaseApp app = FirebaseApp.initializeApp(new FirebaseOptions.Builder()
         .setCredentials(GoogleCredentials.fromStream(ServiceAccount.EDITOR.asStream()))
-        .setFirestoreOptions(FirestoreOptions.newBuilder()
-            .setTimestampsInSnapshotsEnabled(true)
-            .build())
+        .setFirestoreOptions(FIRESTORE_OPTIONS)
         .build());
     Firestore firestore = FirestoreClient.getFirestore(app);
     assertEquals("mock-project-id", firestore.getOptions().getProjectId());
@@ -62,9 +67,7 @@ public class FirestoreClientTest {
     FirebaseApp app = FirebaseApp.initializeApp(new Builder()
         .setCredentials(GoogleCredentials.fromStream(ServiceAccount.EDITOR.asStream()))
         .setProjectId("explicit-project-id")
-        .setFirestoreOptions(FirestoreOptions.newBuilder()
-            .setTimestampsInSnapshotsEnabled(true)
-            .build())
+        .setFirestoreOptions(FIRESTORE_OPTIONS)
         .build());
     Firestore firestore = FirestoreClient.getFirestore(app);
     assertEquals("explicit-project-id", firestore.getOptions().getProjectId());
@@ -104,15 +107,22 @@ public class FirestoreClientTest {
     FirebaseApp app = FirebaseApp.initializeApp(new FirebaseOptions.Builder()
         .setCredentials(GoogleCredentials.fromStream(ServiceAccount.EDITOR.asStream()))
         .setProjectId("mock-project-id")
-        .setFirestoreOptions(FirestoreOptions.newBuilder()
-            .setTimestampsInSnapshotsEnabled(true)
-            .build())
+        .setFirestoreOptions(FIRESTORE_OPTIONS)
         .build());
 
-    assertNotNull(FirestoreClient.getFirestore(app));
+    Firestore firestore = FirestoreClient.getFirestore(app);
+    assertNotNull(firestore);
+    DocumentReference document = firestore.collection("collection").document("doc");
     app.delete();
     try {
       FirestoreClient.getFirestore(app);
+      fail("No error thrown for deleted app");
+    } catch (IllegalStateException expected) {
+      // ignore
+    }
+
+    try {
+      document.get();
       fail("No error thrown for deleted app");
     } catch (IllegalStateException expected) {
       // ignore
@@ -125,5 +135,4 @@ public class FirestoreClientTest {
       // ignore
     }
   }
-
 }